Extract GTModelCMT test method to separate class
